VroVault/requirements.txt
customtkinter>=5.2.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
Pillow>=10.0.0
pyperclip>=1.8.2
zxcvbn>=4.4.28
